# allocate iteration structures lazily