Many times when we go through these functions, we never need to alloc the `seen` or `calls` arrays

@krisselden r?