@krisselden 2017-03-13
looks good, though this whole area could use some more testing.