@stefanpenner 2017-04-04
Benchmarks for this guy:

```js
meta forEachIn
testing
- new (empty)
- old (empty)
- new (notEmpty)
- old (notEmpty)
running first test, please wait...
  new (empty) ...... 57,116,740.98 op/s
  old (empty) ...... 13,495,844.67 op/s
  new (notEmpty) .... 2,384,559.17 op/s
  old (notEmpty) .... 2,346,057.37 op/s
```


For empty traversals (extremely common) we are now 4.3x faster and non empty traversals are about the same.